* lisp/ox-odt.el (ODT_CONTENT_TEMPLATE_FILE, ODT_STYLES_FILE, ODT_EXTRA_IMAGES): Change syntax and/or semantics
* lisp/ox-odt.el (ODT_CONTENT_TEMPLATE_FILE, ODT_STYLES_FILE):
Change syntax.  These are plain strings, instead of
`read'-able strings.
(ODT_EXTRA_IMAGES): When specified with a relative path, if
ODT_STYLES_FILE is a OpenDocument file, the relative path is
expanded against the root of the archive file.  Otherwise, the
path is expanded against the directory of input Org file.

(org-odt-template): Honor above change in syntax and/or
semantics.