* lisp/ox-odt.el: Support for category mapping (from say, a table to figure)
* lisp/ox-odt.el (org-odt-caption-and-numbering-settings):
Replace (a) the variable "Illustration" with "Figure" (b) the
style "Illustration" with "Figure".
(org-odt-category-attribute-to-category): New.  Alist that
maps a `:category' atrribute to category
(org-odt--element-category): Honor `:category' attribute.

An element can forced to be of a certain category by the use of
`:category' property.  For example, the following table will be
captioned as if it is a figure.

    #+NAME: table
    #+CAPTION: A Table
    #+ATTR_ODT: :category "figure"
    | 1 | 2 |
    | 3 | 4 |

This property is particularly useful while typesetting side-by-side
figures using a table like the one below.

    #+NAME: table
    #+CAPTION: Animals
    #+ATTR_ODT: :category "figure"
    #+ATTR_ODT: :list-table t
    -
      -
	#+NAME: dog
	#+CAPTION: A Dog
	[[./dog.png]]
      -
	#+NAME: goat
	#+CAPTION: A Goat
	[[./goat.png]]

See #53.