* lisp/ox-odt.el (org-odt-automatic-styles, org-odt-object-counters): Remove these vars
* lisp/ox-odt.el (org-odt-automatic-styles)
(org-odt-object-counters): : Remove these vars.  Use
`:odt-automatic-styles' and `:odt-object-counters' in the
communication channel.
(org-odt-export-to-odt): Remove references to the above vars.
(org-odt-add-automatic-style): Fold this in to
`org-odt--get-derived-paragraph-style' and `org-odt--table'
and remove it.
(org-odt--get-derived-paragraph-style): Change signature.
Append automatic style, if any, to `:odt-automatic-styles'.
(org-odt-headline, org-odt-paragraph): Use new signature of
`org-odt--get-derived-paragraph-style'.
(org-odt--table): Append automatic style, if any, to
`:odt-automatic-styles'.
(org-odt-template): No need to handle automatic styles for
tables and paragraphs specially.  They are already added to
`:odt-automatic-styles' as and when they are encountered.