* lisp/ox-odt.el:  Implement (i) body-only export and (ii) export to buffer
* lisp/ox-odt.el (org-odt-export-as-odt): New command to
export just the contents as an XML buffer.  It can be invoked
with `C-c C-e o x'.
(org-odt-write-contents-file): Modified.  Now, it handles
writing of the contents to disk.  The assembling of contents
happens now as part of `org-odt-inner-template'.
(org-odt-inner-template): Extracted from
`org-odt-write-contents-file'.  See above.
(org-odt-export-to-odt): Honour body-only export.  When
body-only is non-nil, export just the contents.xml file.
ca
cke
(org-odt-export-to-odt-backend): New.  Extracted from
`org-odt-export-to-odt', and generalized to any OpenDocument
backend.