* lisp/ox-odt.el: Fix file corruption with image/formula in headlines
* lisp/ox-odt.el (org-odt-format-headline--wrap):
`org-export-data-with-backend' works with a new communication
channel that is locally bound.  As a result any state--for
example a manifest entry for an inline image or formula that
is part of the headline--that is stashed in to the
communication channel during the course of the call becomes
unavailable for subsequent calls.  Loss of manifest entries
will make LibreOffice think that the generated ODT file is
"corruption" .  So, use `org-export-data-with-backend' with
caution.

Partial fix for #87.

Here is a sample snippet to test drive this fix

    * An inline image [[./image.png]] in this headline

TODO: Check what happens to TOC entries when there are inline
images / formula in the headline.