* lisp/ox-odt.el (ODT_APP): Preparing for export to Google Documents
* lisp/ox-odt.el: New export export keyword `:odt-app'.  A
`ODT_APP' is the end-user application which will be used to
process the exported document downstream.  For example,
LibreOffice and Google Documents are examples of ODT_APPs.
Experience suggests that an application that is produced for
consumption by LibreOffice may not be suitable for consumption by
other apps like Google Documents.  This is because these other
apps are NOT "full-fledged" OpenDocument apps and support only a
subset of features that is expressible OpenDocument format.  The
value of this keyword, can be used to tailor output to a specific
app.  Currently, the value can be one of nil (="lo") or "gdocs".

As on this date (early Dec. 2021), Google Docs has the following
limitations:

1. No support for caption numbering.  Specifically, no support for
     `text:sequence' and `text:sequence-ref' elements.

2. No support for image caption (as recommended by LibreOffice
Guides).

Accordingly, adapters are needed for typesetting the following:

- captions to images/formulae
- cross-references to captioned image/formulae

(org-odt-format-label, org-odt--render-image/formula): Make these
app-specific.  Use `org-odt--get-app-function', to route it to
app-specific handlers.

(org-odt--get-app-function): New.  Return app-specific handlers for
`render-image/formula' and `format-label'.

(org-odt-format-label-for-lo, org-odt--render-image/formula-for-lo):
LibreOffice-specific handlers for `render-image/formula' and
`format-label'.  These are verbatim copies of current
`org-odt-format-label' and `org-odt--render-image/formula'.
(org-odt-format-label-for-gdocs,

org-odt--render-image/formula-for-gdocs): 'Google Documents'-specific
handlers for `render-image/formula' and `format-label'.  These are
verbatim copies of current `org-odt-format-label' and
`org-odt--render-image/formula'.  In other words, even though `Google
Documents'-specific handlers are in place, the specifics aren't taken
care of yet.

(org-odt-link--inline-image, org-odt-link--inline-formula): Introduce
new ODT attribute `:app' for captioned images/formulae.  Have
`org-odt-link--inline-formula' ignore the document-wide and
element-specfic app value.

Preparing for export to 'Google Documents'.  See #136, #137 and #145.