* lisp/ox-odt.el (ODT_APP):  Support for non-LibreOffice apps
* lisp/ox-odt.el (ODT_APP): Specify the ODT consumer.  When it is
nil, or string "lo", assume that the ultimate consumer is
LibreOffice.  When it is other than these values, assume that the
consumer is either Microsoft Word or Google Docs.
(org-odt-link--inline-image): (a) Don't use "paragraph" anchoring.
Instead use only "as-char" and "char" anchoring.  (b) In case of
non-LibreOffice apps, don't generate an "outer frame".  These
ensure that the ODT file created this by exporter, when
transformed to Microsoft Word format, or imported in to Google
Docs doesn't suffer any "distortions".
(org-odt-link--inline-formula): Change signature.  Behaves the
same, irrespective of the value of `:odt-app'.

Partial fix for #153, and issue discussed in
discussions/170#discussioncomment-2369765.