* lisp/ox-odt.el: Use a 2-column table to typeset a formula and its number
* etc/styles/OrgOdtContentTemplate.xml (OrgEquationTable)
(OrgEquationTableColumn1, OrgEquationTableColumn2): New automatic
styles.

* etc/styles/OrgOdtStyles.xml (OrgToCharFormula)
(OrgToCharFormulaCaptionAbove): New styles.  For future use, may
be.
* lisp/ox-odt.el (org-odt-link--inline-formula): Use a 2-column
table to typeset a formula and its number.
(org-odt-link--inline-formula): Always use `as-char' as anchor for
typesetting formula.
(org-odt--render-formula): Use line break to separate caption and
its formula, for now.
(org-odt-paragraph): Skip emitting a paragraph when the contents
is a formula.