* lisp/ox-ods.el: Support for `$#' and `@#', and locale specific behaviour
* lisp/ox-ods.el (org-ods-cell-mapper): Set this to
`org-ods-replace-org-ts-with-locale-ts-in-string'.  So, that
brackets around timestamps are stripped, and the timestamp is
converted in to locale specific string.  If this is not what you
want, set this to `nil'.
(org-ods-calc-f->ods-f-alist): Add entries for `vmean', `$#', and `@#'.
(org-ods-convert, org-ods-translate): If you bind
`TABLE_CONVERTER_ENV' as below,

    `#+BIND: TABLE_CONVERTER_ENV (("LANG" "de_DE.UTF-8"))',

then the `org-odt-convert' and
`org-ods-translate' behave in locale-specific way.
(org-ods--with-environment): New macro.

* lisp/ox-odt.el (org-odt-convert-processes): Add Gnumeric/ssconvert
(org-odt-do-convert): Add `%l' format-spec for passing on a LANG
environment variable.

More fixes for #197.