* lisp/ox-ods.el (org-ods-convert-process): New user option
* lisp/ox-ods.el (org-ods-convert-process): Converter name to use
for converting ODS files to other related formats.  Defaults to
"LibreOffice".

(org-ods-export-to-ods): Let bind `org-odt-convert-process' to
`org-ods-convert-process'.

This changes enables having different converters for ODT and ODS
documents.  For example, one may want `soffice'/`LibreOffice' as a
converter for `odt' documents, but want `ssconvert'/`Gnumeric' as
a converter for ODS documents.

See #197.