Fix issue where some styles in style files are overridden
When a style file specified using the `ODT_STYLES_FILE` key includes
a style identically named to one of the generated Htmlfontify styles,
the latter overrides the former.

The fix is to put generated Htmlfontify styles at the beginning of the
`<office:styles>` tag contents instead of the end. This way the
`ODT_STYLES_FILE` will take precidence.

Fixes issue #215.