* lisp/odt-continued.el: Manipulate OpenDocument XML as Lisp Object (= DOM)
* lisp/odt-continued.el: New library for working with XML DOMs in
general, OpenDocument DOMs in particuar.

See below for the main work-horse functions of this library.  Be
warned that none of the below entry points are "correct" or
"complete".  The entry points also differ in their "maturity".  In
other words, this library is contains prototype code and is a
work-in-progress.

The ultimate aim of this library to build EIEIO classes out of
OD elements, especifically the OD style elements, and use EIEIO setters /
getters / cloners to manipulate style definitions as Lisp objects.  To
this end, this library contains

 (1) an RNG -> RNC converter
 (2) APIs for class builders
 (3) a prototype class buider

(odt-rngdom:rng->rnc): Convert an RNG file to RNC file.  Right now
it can convert the official `OpenDocument-v1.2-os-schema.rng' to a
rnc file that is as good as the `OpenDocument-v1.2-os-schema.rnc'
file produced by the `trang' utility.  See #225.

(odt-stylesdom:inspect-type): Given a OD style instance, infer its
parent, attribute and child tags.  This is a quick and dirty way
to build EIEIO classes for OD style objects.  Contrast this
"dirty" approach with a much "cleaner" and "comprehensive"
approach of relying on definitions in the OD RNG files to building
the EIEIO classes.  See #222.

(odt-class:type-instance-alist->defclasses): Convert an OD style
instance to a `defclass'.  See #222.

(odt-class:type-forms-alist->file): Persist lisp forms to a `*.el'
file.  See #222.

(odt-class:unit-test): Running this function creates a
`odt-classes.el' file.  See `odt-classes.el' for how a style may
be represented as a EIEIO class.  See #222.

(odt-contentsdom:contents.xml->org-file): Convert a `contents.xml'
(or equivalently text within `<office:document-content>...</office:document-content>') to
an `Org' file.  See #223.