* lisp/ox-odt.el (org-odt--table): New ODT attribute `:col-width'
* lisp/ox-odt.el (org-odt--table, org-odt--table-cell-widths):
Modified. Add support for new ODT attribute `:col-width'.

Fixes #245.

Add support for new odt attribute `:col-width'.  The value is a float,
and it is interpreted in `cm' (=centi-metre) units.

In other words, the actual width of a column is the column's width
value as determined from table's `:widths' attribute multiplied by
this value.

Right now,

- `:col-width' is checked only for ODS tables
- when `:col-width' is specified, the `:rel-width' attribute has
no effect.

Here is a test snippet to test drive this change:

    #+ATTR_ODT: :style "GriddedTable"
    #+ATTR_ODT: :col-width 2.5
    #+CAPTION: simple
    | Width of this column in 2.5cm | and width of this column is 2.5cm |

    #+ATTR_ODT: :style "GriddedTable"
    #+ATTR_ODT: :col-width 2.5
    #+ATTR_ODT: :widths "2,3"
    #+CAPTION: simple
    | Width of this column is 5cm (= 2 * 2.5cm) | while the width of this column is 7.5cm  (= 3 * 2.5cm) |