* lisp/ox-ods.el (org-ods-formula-arg-separator): New variable
* lisp/ox-odt.el (org-odt-table-cell): When the `Org' table cell
is empty, ensure that the OpenDocument Table Cell is also empty.
This ensures that `ISBLANK' is TRUE for that cell.

* lisp/ox-ods.el (org-ods-formula-arg-separator): New variable.
Its value contros the character that is used as separator in ODS
formula.  It can either be comma (=`,') or semi-colon (=`;')
depending upon the locale.

(org-ods-insert-ods-formula): Replace comma (=`,') with

`org-ods-formula-arg-separator'.  The replacement is a very dumb
replacement.

(org-ods-table-cell): Fix parsing of numbers in table cell.  The
formula cell has no attribute other than the ODS formula.
Specifically its value attribute is empty.  This means that one
needs to rely on LibreOffice's recalculate functionality accessed
with `Data' -> `Calculate' -> `Recalculate Hard' (= Shift + Ctrl +
F9) to populate formula cells.

Experimental changes for #252.

Here is a sample `Org' table to explore this change.

    #+begin_src emacs-lisp :exports results :results none
      (setq org-ods-formula-arg-separator ";")

      (defmath ISBLANK (x)
	(if (numberp x)
	    0
	  1))
    #+end_src

    #+Name: test-if
    | Miller | pos | range | weight |
    |--------+-----+-------+--------|
    |        |  33 |   329 |  20297 |
    |        |  71 |   329 |  20259 |
    |        |  95 |   267 |  20173 |
    |        | 102 |   329 |  20228 |
    |        |  64 |   333 |  20270 |
    |        |  48 |   329 |  20282 |
    |        |  55 |   329 |  20275 |
    |        | 124 |   329 |  20206 |
    |        |     |       |      0 |
    |        |     |       |      0 |
    |        |     |       |      0 |
    |        |     |       |      0 |
    |--------+-----+-------+--------|
    |        |     |       | 161990 |
    #+TBLFM: $4=if(ISBLANK($3),0,30000+($3+1)-$2);E f2::@14$4=vsum(@I$4..@II$4);f2
    # #+TBLFM: $4=if("$3"=="nan",string(""),($3+1)-$2);E f2::@14$4=vsum(@I$4..@II$4);f2