* lisp/odt.el: (pcase--flip-3ary): New macro; replaces `map--pcase-map-elt'
* lisp/odt.el (odt-map--make-pcase-bindings): Use it.

Note that `map--pcase-map-elt' is not available on Emacs 29.1, the
latest released version, but is available only on Emacs 30, the
unreleased development version.  The above fix allows compilation
against Emacs 29.

If there are non-trivial build breakages with older versions of Emacs,
because of non-availability required symbols, it is better to use
cutting-edge libraries from `gnu-devel' ELPA repository at
https://elpa.gnu.org/devel/.

Fixes #263.