* lisp/ox-odt.el (org-odt-yank-styles, org-odt-insert-style-name-or-style-definition-from-file): Enhanced to specify styles for ODS backend
* lisp/ox-odt.el (org-odt-supported-file-types): Add `ods'.

(org-odt-supported-backends, org-odt-backend-and-file-types): New
constants.

(org-odt-dom-from-a-styles-file): New.  Workhorse function that
prompts for a styles file, and returns XML Dom.

(org-odt-yank-styles): Modified to support the all new `ods'
backend.

Add a new parameter `which-backend'.  Remove the `odt' prefix from
`which-style' param to drive home the notion that the function can
be used to yank any backend style.

Allow `ods' and `ots' extensions while prompting for a
styles-file.

Enhance the `#+ATTR_ODT: ' line of a `nxml' src block, to also
take a `:backends' property in addition to the existing `:targets'
property.  Value of `:backends' is a list of backends (as symbols)
or a symbol `any'.  The `:backends' value specifies the backends
in to which this style block has to be included.

Generate backend-specific keywords style keywords rather than a
odt-specific keyword.

(org-odt-insert-style-name-or-style-definition-from-file):
Enhanced to insert style from ODS, OTS backend as well.  See
previous comment for details.

----------------

Add support for ODS-specific style blocks in addition to the existing
ODT-specific style blocks.

Preparing to address #266.